Fixed the test cases and fixed the IPv6 version
The reason the IPv6 cases failed was that we created a net.IP object and then called the string method on that. The string method truncates the leading zeroes in a block. Just using the original string that was used to check if it is a valid IP adres return the original C++ version's behaviour.
The test cases are output from the C++ version using the provided script.